Issue #11163: Enforce file size InputFinalLocalVariable2 java file word 'initialisation' fixed to 'initialization'
