+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/coding/messages_fr.properties
@@ -1,7 +1,7 @@
 array.trailing.comma=La dernière valeur d''un tableau devrait être suivie d''une virgule.
 assignment.inner.avoid=Évitez d''affecter une valeur à une variable au sein d''une expression.
 avoid.clone.method=Évitez d''utiliser la méthode de clonage.
-avoid.double.brace.init=Évitez l''initialisation à double accolade.
+avoid.double.brace.init=Évitez l''initialization à double accolade.
 avoid.finalizer.method=Évitez d''utiliser la méthode de finalisation.
 covariant.equals=Votre méthode equals compare uniquement les objets de votre classe. N''oubliez pas de surcharger la méthode equals(java.lang.Object).
 declaration.order.access=La définition des variables n''est pas triée suivant leur portée.
@@ -15,7 +15,7 @@ equals.avoid.null=Les chaînes littérales devraient être sur le côté gauche 
 equals.noEquals=La définition de la méthode ''hashCode()'' doit toujours être accompagnée de la définition de la méthode ''equals()''.
 equals.noHashCode=La définition de la méthode ''equals()'' doit toujours être accompagnée de la définition de la méthode ''hashCode()''.
 equalsIgnoreCase.avoid.null=Les chaînes littérales devraient être sur le côté gauche d''une comparaison ''equalsIgnoreCase''.
-explicit.init=L''initialisation explicite de la variable ''{0}'' à la valeur ''{1}'' est inutile, c''est la valeur par défaut pour ce type.
+explicit.init=L''initialization explicite de la variable ''{0}'' à la valeur ''{1}'' est inutile, c''est la valeur par défaut pour ce type.
 fall.through=Le cas précédent du "switch" ne contient pas de break, return, yield, throw ou continue.
 fall.through.last=Le dernier cas du "switch" ne contient pas de break, return, yield, throw ou continue.
 final.variable=La variable ''{0}'' devrait être finale.
