+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/design/messages_fr.properties
@@ -1,4 +1,4 @@
-arrangement.members.before.inner=Les blocs d''initialisation, les constructeurs, \
+arrangement.members.before.inner=Les blocs d''initialization, les constructeurs, \
   les champs et les méthodes doivent être avant les types internes.
 design.forExtension=La classe ''{0}'' semble être conçue pour extension (peut être héritée), mais la méthode\
   ''{1}'' n''a pas de Javadoc qui explique comment le faire en toute sécurité. Si la classe n''est pas \
